move job impl to separate classa nd raml file
