+++ ramls/oai-pmh.raml
@@ -93,25 +93,3 @@ resourceTypes:
               example:
                 strict: false
                 value: !include examples/service_unavailable.sample
-  /clean-up-instances:
-        displayName: Clean up data for outdated requests from instances table
-        description: API to start clean up mechanism of retained records from instances table
-        post:
-          is: [validate]
-          responses:
-            204:
-            400:
-              description: "Bad request"
-              body:
-                text/plain:
-                  example: "Bad request"
-            422:
-              description: "Unprocessable Entity"
-              body:
-                application/json:
-                  type: errors
-            500:
-              description: "Internal server error"
-              body:
-                text/plain:
-                  example: "Internal server error"
